Refactored token prices chart
